feature: 별점 자료형 수정

+++ src/main/java/com/trip/aslung/review/model/dto/PostScheduleDto.java
@@ -9,7 +9,7 @@ public class PostScheduleDto {
     private String placeName; // 장소명
     private String memo;      // 메모 (설명)
     private String category;  // 카테고리
-    private int rating;  // 장소별 리뷰 평점
+    private double rating;  // 장소별 리뷰 평점
 
     // [추가] 리뷰 코멘트 필드
     private String comment;
